requirements.txt
streamlit
torch
transformers
opencv-python-headless
pillow
numpy
sentencepiece
protobuf